Skip some tests for token addition behavior
