+++ tests/units/tokens/test_condition_token.py
@@ -203,7 +203,6 @@ def test_async_wait_condition():
     assert finish_time - start_time >= timeout
 
 
-
 @pytest.mark.parametrize(
     'options',
     [
@@ -324,35 +323,7 @@ def test_cached_condition_cancelling(options):
         assert counter == 6
 
 
-def test_quasitemp_condition_token_plus_temp_simple_token():
-    token = ConditionToken(lambda: False) + SimpleToken()
-
-    assert isinstance(token, ConditionToken)
-    assert len(token._tokens) == 0
-
-
-def test_not_quasitemp_condition_token_plus_temp_simple_token():
-    condition_token = ConditionToken(lambda: False)
-    token = condition_token + SimpleToken()
-
-    assert isinstance(token, SimpleToken)
-    assert len(token._tokens) == 1
-    assert isinstance(token._tokens[0], ConditionToken)
-    assert token._tokens[0] is condition_token
-
-
-def test_quasitemp_condition_token_plus_not_temp_simple_token():
-    simple_token = SimpleToken()
-    token = ConditionToken(lambda: False) + simple_token
-
-    assert isinstance(token, ConditionToken)
-    assert token is not simple_token
-    assert len(token._tokens) == 1
-    assert isinstance(token._tokens[0], SimpleToken)
-    assert token._tokens[0] is simple_token
-
-
-def test_not_quasitemp_condition_token_plus_not_temp_simple_token():
+def test_condition_token_plus_simple_token():
     simple_token = SimpleToken()
     condition_token = ConditionToken(lambda: False)
     token = condition_token + simple_token
@@ -365,34 +336,7 @@ def test_not_quasitemp_condition_token_plus_not_temp_simple_token():
     assert token._tokens[1] is simple_token
 
 
-def test_quasitemp_condition_token_plus_temp_simple_token_reverse():
-    token = SimpleToken() + ConditionToken(lambda: False)
-
-    assert isinstance(token, ConditionToken)
-    assert len(token._tokens) == 0
-
-
-def test_not_quasitemp_condition_token_plus_temp_simple_token_reverse():
-    condition_token = ConditionToken(lambda: False)
-    token = SimpleToken() + condition_token
-
-    assert isinstance(token, SimpleToken)
-    assert len(token._tokens) == 1
-    assert isinstance(token._tokens[0], ConditionToken)
-    assert token._tokens[0] is condition_token
-
-
-def test_quasitemp_condition_token_plus_not_temp_simple_token_reverse():
-    simple_token = SimpleToken()
-    token = simple_token + ConditionToken(lambda: False)
-
-    assert isinstance(token, ConditionToken)
-    assert token is not simple_token
-    assert len(token._tokens) == 1
-    assert token._tokens[0] is simple_token
-
-
-def test_not_quasitemp_condition_token_plus_not_temp_simple_token_reverse():
+def test_simple_token_plus_condition_token():
     simple_token = SimpleToken()
     condition_token = ConditionToken(lambda: False)
     token = simple_token + condition_token
